requirements.txt
gradio>=4.0
python-dotenv>=1.0
openai~=1.98.0
fastapi>=0.116.1
uvicorn>=0.35.0
requests>=2.32.4

pytest>=8.0

